[Fix] Fixed the code
The status was initally returning an error 404 for the route
curl -XGET http://0.0.0.0:5000/api/v1/status

This was because the root route was set to /api/v1. so the correct route
would have been curl -XGET http://0.0.0.0:5000/api/vi/api/v1/status
which is not the right route. To fix the issue, I changed it to "/"